+++ Sources/iOS Utils/DesignableButton.swift
@@ -1,8 +1,6 @@
-#if canImport(UIKit)
-
 import UIKit
 
-@available(iOS 11.0, *)
+@available(iOS 11.0, tvOS 11.0, *)
 @IBDesignable
 public class DesignableButton: UIButton {
     // MARK: - Corners
@@ -162,5 +160,3 @@ public class DesignableButton: UIButton {
         layer.shouldRasterize = shadowColor != .clear && shadowOpacity != 0.0
     }
 }
-
-#endif