+++ Sources/iOS Utils/UIAlertController+Convenience.swift
@@ -1,5 +1,3 @@
-#if canImport(UIKit)
-
 import UIKit
 
 public typealias AlertActionHandler = (UIAlertAction) -> Void
@@ -106,5 +104,3 @@ public extension UIAlertController {
         alertActions.forEach { addAction($0.alertAction) }
     }
 }
-
-#endif