+++ Sources/iOS Utils/UIApplication+UIWindow.swift
@@ -1,5 +1,3 @@
-#if canImport(UIKit)
-
 import UIKit
 
 extension UIApplication {
@@ -13,5 +11,3 @@ extension UIApplication {
         }
     }
 }
-
-#endif